+++ RealmSwift/Query.swift
@@ -21,7 +21,7 @@ import Realm
 import Realm.Private
 
 /// Enum representing an option for `String` queries.
-public struct StringOptions: OptionSet {
+public struct StringOptions: OptionSet, Sendable {
     /// :doc:
     public let rawValue: Int8
     /// :doc: